moved function to compare plugin
